Revert to passing instance (rather than class) to MultiprocessInput.
This feels unsatisfactory, but some properties can only be properly
known once the device has been instantiated, esp. when the size of the
expected data is some function of user input (e.g. if there are two
flock of birds, size of data is 6; if three, then 9).

This is unsatisfactory because it may lead to confusion w.r.t. which
instantiation to use (use the device vs use the MultiprocessInput), and
may lead to unexpected behavior if the device is first used locally,
then used on the remote process (e.g. the keyboard module is very
unhappy about touching multiple processes).