Add testing device, update HAND, keyboard, and mouse with LSL.
Note that the keyboard implementation does not yet return the value of
the pressed key, just the index (relative to the subset of keys provided
to the constructor). There was a memory leak in lsl/pylsl when dealing
with strings, and that fix is not present in the pypi release yet.